requirements.txt
numpy>=1.20.0
scikit-learn>=1.0.0
tomli>=2.0.0
pyyaml>=6.0
torch>=1.12.0
transformers>=4.25.0